+++ src/lazy-collection.ts
@@ -107,12 +107,21 @@ export class LazyCollection<T> implements ICollection<T> {
   }
 
   /**
-   * Fork into nested collections
-   * Inner collections are grouped by their return from the emit fn
+   * Exclude falsy values from the collection
+   *
+   * @param this
+   * @returns
+   */
+  compact(): LazyCollection<NonNullable<T>> {
+    return this.filter(Boolean) as $TODO<$ANY>;
+  }
+
+  /**
+   * Partition the collection grouping by callback return value
    *
    * @param forks
    */
-  forkOn<R>(callbackfn: ((value: T, index: number) => R)): LazyCollection<LazyCollection<T>> {
+  partition<R>(callbackfn: ((value: T, index: number) => R)): LazyCollection<LazyCollection<T>> {
     const self = this;
     function * Iterateable(): IterableIterator<LazyCollection<T>> {
       // need to collect all results to group them
@@ -139,7 +148,7 @@ export class LazyCollection<T> implements ICollection<T> {
   forkFlat<R>(...forks: readonly (Unary<this, Iterateable<R>>)[]): LazyCollection<R> {
     const combined = this
       // split forks
-      .forkMap(...forks)
+      .all(...forks)
       .flat()
       // normalize
       .map((splits) => new LazyCollection(splits))
@@ -153,27 +162,27 @@ export class LazyCollection<T> implements ICollection<T> {
    *
    * Similar to Promise.all
    */
-  forkMap<M extends Record<PropertyKey, Unary<this, unknown>>>(forks: M): LazyCollection<{ [K in keyof M]: ReturnType<M[K]> }>;
-  forkMap<R1>(...forks: readonly [Unary<this, R1>]): LazyCollection<[R1]>
-  forkMap<R1, R2>(...forks: readonly [Unary<this, R1>, Unary<this, R2>]): LazyCollection<[R1, R2]>
-  forkMap<R1, R2, R3>(...forks: readonly [Unary<this, R1>, Unary<this, R2>, Unary<this, R3>]): LazyCollection<[R1, R2, R3]>
-  forkMap<R1, R2, R3, R4>(...forks: readonly [Unary<this, R1>, Unary<this, R2>, Unary<this, R3>, Unary<this, R4>]): LazyCollection<[R1, R2, R3, R4]>
-  forkMap<R1, R2, R3, R4, R5>(...forks: readonly [Unary<this, R1>, Unary<this, R2>, Unary<this, R3>, Unary<this, R4>, Unary<this, R5>]): LazyCollection<[R1, R2, R3, R4, R5]>
-  forkMap<R1, R2, R3, R4, R5, R6>(...forks: readonly [Unary<this, R1>, Unary<this, R2>, Unary<this, R3>, Unary<this, R4>, Unary<this, R5>, Unary<this, R6>]): LazyCollection<[R1, R2, R3, R4, R5, R6]>
-  forkMap<R1, R2, R3, R4, R5, R6, R7>(...forks: readonly [Unary<this, R1>, Unary<this, R2>, Unary<this, R3>, Unary<this, R4>, Unary<this, R5>, Unary<this, R6>, Unary<this, R7>]): LazyCollection<[R1, R2, R3, R4, R5, R6, R7]>
-  forkMap<R1, R2, R3, R4, R5, R6, R7, R8>(...forks: readonly [Unary<this, R1>, Unary<this, R2>, Unary<this, R3>, Unary<this, R4>, Unary<this, R5>, Unary<this, R6>, Unary<this, R7>, Unary<this, R8>]): LazyCollection<[R1, R2, R3, R4, R5, R6, R7, R8]>
-  forkMap<R1, R2, R3, R4, R5, R6, R7, R8, R9>(...forks: readonly [Unary<this, R1>, Unary<this, R2>, Unary<this, R3>, Unary<this, R4>, Unary<this, R5>, Unary<this, R6>, Unary<this, R7>, Unary<this, R8>, Unary<this, R9>]): LazyCollection<[R1, R2, R3, R4, R5, R6, R7, R8, R9]>
-  forkMap<R1, R2, R3, R4, R5, R6, R7, R8, R9, R10>(...forks: readonly [Unary<this, R1>, Unary<this, R2>, Unary<this, R3>, Unary<this, R4>, Unary<this, R5>, Unary<this, R6>, Unary<this, R7>, Unary<this, R8>, Unary<this, R9>, Unary<this, R10>]): LazyCollection<[R1, R2, R3, R4, R5, R6, R7, R8, R9, R10]>
-  forkMap<R1, R2, R3, R4, R5, R6, R7, R8, R9, R10, R11>(...forks: readonly [Unary<this, R1>, Unary<this, R2>, Unary<this, R3>, Unary<this, R4>, Unary<this, R5>, Unary<this, R6>, Unary<this, R7>, Unary<this, R8>, Unary<this, R9>, Unary<this, R10>, Unary<this, R11>]): LazyCollection<[R1, R2, R3, R4, R5, R6, R7, R8, R9, R10, R11]>
-  forkMap<R1, R2, R3, R4, R5, R6, R7, R8, R9, R10, R11, R12>(...forks: readonly [Unary<this, R1>, Unary<this, R2>, Unary<this, R3>, Unary<this, R4>, Unary<this, R5>, Unary<this, R6>, Unary<this, R7>, Unary<this, R8>, Unary<this, R9>, Unary<this, R10>, Unary<this, R11>, Unary<this, R12>]): LazyCollection<[R1, R2, R3, R4, R5, R6, R7, R8, R9, R10, R11, R12]>
-  forkMap<R1, R2, R3, R4, R5, R6, R7, R8, R9, R10, R11, R12, R13>(...forks: readonly [Unary<this, R1>, Unary<this, R2>, Unary<this, R3>, Unary<this, R4>, Unary<this, R5>, Unary<this, R6>, Unary<this, R7>, Unary<this, R8>, Unary<this, R9>, Unary<this, R10>, Unary<this, R11>, Unary<this, R12>, Unary<this, R13>]): LazyCollection<[R1, R2, R3, R4, R5, R6, R7, R8, R9, R10, R11, R12, R13]>
-  forkMap<R1, R2, R3, R4, R5, R6, R7, R8, R9, R10, R11, R12, R13, R14>(...forks: readonly [Unary<this, R1>, Unary<this, R2>, Unary<this, R3>, Unary<this, R4>, Unary<this, R5>, Unary<this, R6>, Unary<this, R7>, Unary<this, R8>, Unary<this, R9>, Unary<this, R10>, Unary<this, R11>, Unary<this, R12>, Unary<this, R13>, Unary<this, R14>]): LazyCollection<[R1, R2, R3, R4, R5, R6, R7, R8, R9, R10, R11, R12, R13, R14]>
-  forkMap<R1, R2, R3, R4, R5, R6, R7, R8, R9, R10, R11, R12, R13, R14, R15>(...forks: readonly [Unary<this, R1>, Unary<this, R2>, Unary<this, R3>, Unary<this, R4>, Unary<this, R5>, Unary<this, R6>, Unary<this, R7>, Unary<this, R8>, Unary<this, R9>, Unary<this, R10>, Unary<this, R11>, Unary<this, R12>, Unary<this, R13>, Unary<this, R14>, Unary<this, R15>]): LazyCollection<[R1, R2, R3, R4, R5, R6, R7, R8, R9, R10, R11, R12, R13, R14, R15]>
-  forkMap<R>(...forks: readonly (Unary<this, R>)[]): LazyCollection<R[]>
-  forkMap(mapOrFunction?: (Unary<this, unknown> | Record<PropertyKey, Unary<this, unknown>>), ...rest: Unary<this, unknown>[]): $ANY {
+  all<M extends Record<PropertyKey, Unary<this, unknown>>>(forks: M): LazyCollection<{ [K in keyof M]: ReturnType<M[K]> }>;
+  all<R1>(...splits: readonly [Unary<this, R1>]): LazyCollection<[R1]>
+  all<R1, R2>(...splits: readonly [Unary<this, R1>, Unary<this, R2>]): LazyCollection<[R1, R2]>
+  all<R1, R2, R3>(...splits: readonly [Unary<this, R1>, Unary<this, R2>, Unary<this, R3>]): LazyCollection<[R1, R2, R3]>
+  all<R1, R2, R3, R4>(...splits: readonly [Unary<this, R1>, Unary<this, R2>, Unary<this, R3>, Unary<this, R4>]): LazyCollection<[R1, R2, R3, R4]>
+  all<R1, R2, R3, R4, R5>(...splits: readonly [Unary<this, R1>, Unary<this, R2>, Unary<this, R3>, Unary<this, R4>, Unary<this, R5>]): LazyCollection<[R1, R2, R3, R4, R5]>
+  all<R1, R2, R3, R4, R5, R6>(...splits: readonly [Unary<this, R1>, Unary<this, R2>, Unary<this, R3>, Unary<this, R4>, Unary<this, R5>, Unary<this, R6>]): LazyCollection<[R1, R2, R3, R4, R5, R6]>
+  all<R1, R2, R3, R4, R5, R6, R7>(...splits: readonly [Unary<this, R1>, Unary<this, R2>, Unary<this, R3>, Unary<this, R4>, Unary<this, R5>, Unary<this, R6>, Unary<this, R7>]): LazyCollection<[R1, R2, R3, R4, R5, R6, R7]>
+  all<R1, R2, R3, R4, R5, R6, R7, R8>(...splits: readonly [Unary<this, R1>, Unary<this, R2>, Unary<this, R3>, Unary<this, R4>, Unary<this, R5>, Unary<this, R6>, Unary<this, R7>, Unary<this, R8>]): LazyCollection<[R1, R2, R3, R4, R5, R6, R7, R8]>
+  all<R1, R2, R3, R4, R5, R6, R7, R8, R9>(...splits: readonly [Unary<this, R1>, Unary<this, R2>, Unary<this, R3>, Unary<this, R4>, Unary<this, R5>, Unary<this, R6>, Unary<this, R7>, Unary<this, R8>, Unary<this, R9>]): LazyCollection<[R1, R2, R3, R4, R5, R6, R7, R8, R9]>
+  all<R1, R2, R3, R4, R5, R6, R7, R8, R9, R10>(...splits: readonly [Unary<this, R1>, Unary<this, R2>, Unary<this, R3>, Unary<this, R4>, Unary<this, R5>, Unary<this, R6>, Unary<this, R7>, Unary<this, R8>, Unary<this, R9>, Unary<this, R10>]): LazyCollection<[R1, R2, R3, R4, R5, R6, R7, R8, R9, R10]>
+  all<R1, R2, R3, R4, R5, R6, R7, R8, R9, R10, R11>(...splits: readonly [Unary<this, R1>, Unary<this, R2>, Unary<this, R3>, Unary<this, R4>, Unary<this, R5>, Unary<this, R6>, Unary<this, R7>, Unary<this, R8>, Unary<this, R9>, Unary<this, R10>, Unary<this, R11>]): LazyCollection<[R1, R2, R3, R4, R5, R6, R7, R8, R9, R10, R11]>
+  all<R1, R2, R3, R4, R5, R6, R7, R8, R9, R10, R11, R12>(...splits: readonly [Unary<this, R1>, Unary<this, R2>, Unary<this, R3>, Unary<this, R4>, Unary<this, R5>, Unary<this, R6>, Unary<this, R7>, Unary<this, R8>, Unary<this, R9>, Unary<this, R10>, Unary<this, R11>, Unary<this, R12>]): LazyCollection<[R1, R2, R3, R4, R5, R6, R7, R8, R9, R10, R11, R12]>
+  all<R1, R2, R3, R4, R5, R6, R7, R8, R9, R10, R11, R12, R13>(...splits: readonly [Unary<this, R1>, Unary<this, R2>, Unary<this, R3>, Unary<this, R4>, Unary<this, R5>, Unary<this, R6>, Unary<this, R7>, Unary<this, R8>, Unary<this, R9>, Unary<this, R10>, Unary<this, R11>, Unary<this, R12>, Unary<this, R13>]): LazyCollection<[R1, R2, R3, R4, R5, R6, R7, R8, R9, R10, R11, R12, R13]>
+  all<R1, R2, R3, R4, R5, R6, R7, R8, R9, R10, R11, R12, R13, R14>(...splits: readonly [Unary<this, R1>, Unary<this, R2>, Unary<this, R3>, Unary<this, R4>, Unary<this, R5>, Unary<this, R6>, Unary<this, R7>, Unary<this, R8>, Unary<this, R9>, Unary<this, R10>, Unary<this, R11>, Unary<this, R12>, Unary<this, R13>, Unary<this, R14>]): LazyCollection<[R1, R2, R3, R4, R5, R6, R7, R8, R9, R10, R11, R12, R13, R14]>
+  all<R1, R2, R3, R4, R5, R6, R7, R8, R9, R10, R11, R12, R13, R14, R15>(...splits: readonly [Unary<this, R1>, Unary<this, R2>, Unary<this, R3>, Unary<this, R4>, Unary<this, R5>, Unary<this, R6>, Unary<this, R7>, Unary<this, R8>, Unary<this, R9>, Unary<this, R10>, Unary<this, R11>, Unary<this, R12>, Unary<this, R13>, Unary<this, R14>, Unary<this, R15>]): LazyCollection<[R1, R2, R3, R4, R5, R6, R7, R8, R9, R10, R11, R12, R13, R14, R15]>
+  all<R>(...splits: readonly (Unary<this, R>)[]): LazyCollection<R[]>
+  all(mapOrFunction?: (Unary<this, unknown> | Record<PropertyKey, Unary<this, unknown>>), ...rest: Unary<this, unknown>[]): $ANY {
 
     if (!mapOrFunction)
-      throw new TypeError('LazyCollection.forkMap: you must provide an object or at least one function');
+      throw new TypeError('[@nkp/iterable::LazyCollection.all]: you must provide an object or at least one function');
 
     const self = this;
     if (typeof mapOrFunction === 'object') {
@@ -230,6 +239,16 @@ export class LazyCollection<T> implements ICollection<T> {
       }
     }
     return new LazyCollection(iterable);
+  }
+
+  /**
+   * Map this instance
+   *
+   * @param callbackfn
+   * @returns
+   */
+  mapSelf<U>(callbackfn: (self: this) => U): U {
+    return callbackfn(this);
   }
 
   /**
@@ -335,7 +354,7 @@ export class LazyCollection<T> implements ICollection<T> {
    *
    * @param value
    */
-  exclude( ...values: T[]): LazyCollection<T> {
+  exclude( ...values: readonly T[]): LazyCollection<T> {
     if (values.length === 0) {
       return new LazyCollection(this);
     }
@@ -415,7 +434,7 @@ export class LazyCollection<T> implements ICollection<T> {
    *
    * @param value
    */
-  pick( ...values: T[]): LazyCollection<T> {
+  pick( ...values: readonly T[]): LazyCollection<T> {
     if (values.length === 0) {
       return new LazyCollection([]);
     }
@@ -520,7 +539,7 @@ export class LazyCollection<T> implements ICollection<T> {
    *
    * @returns
    */
-  push(...pushed: T[]): LazyCollection<T> {
+  push(...pushed: readonly T[]): LazyCollection<T> {
     const self = this;
     function * iteratorable (): Iterable<T> {
       yield * self;
@@ -543,7 +562,7 @@ export class LazyCollection<T> implements ICollection<T> {
    *
    * @returns
    */
-  unshift(...unshifted: T[]): LazyCollection<T> {
+  unshift(...unshifted: readonly T[]): LazyCollection<T> {
     const self = this;
     function * iteratorable (): Iterable<T> {
       yield * unshifted;
@@ -654,7 +673,7 @@ export class LazyCollection<T> implements ICollection<T> {
    * @param right
    * @returns
    */
-  zipShort<U>(right: Iterable<U>): LazyCollection<[T, U]> {
+  zip<U>(right: Iterable<U>): LazyCollection<[T, U]> {
     const self = this;
     function * iteratorable (): Iterable<[T, U]> {
       const _left = self[Symbol.iterator]();
@@ -1013,3 +1032,4 @@ export class LazyCollection<T> implements ICollection<T> {
     return `[object ${this.constructor.name} (${this.getSize()}) { ${display.join(', ')}... }]`;
   }
 }
+