+++ src/test/not-nullable.spec.util.ts
@@ -5,16 +5,16 @@ export function testNotNullable(create: <T>(iterable: Iterateable<T>) => IHasNot
   describe('notNullable(...)', () => {
     describe('should work on', () => {
       it('array', () => {
-        const pipeline = create([1, null, undefined, 3,]);
-        const called: number[] = Array.from(pipeline
+        const collection = create([1, null, undefined, 3,]);
+        const called: number[] = Array.from(collection
           .notNullable()
         );
         expect(called[0]).toEqual(1);
         expect(called[1]).toEqual(3);
       });
       it('set', () => {
-        const pipeline = create(new Set([1, null, undefined, 3,]));
-        const called: number[] = Array.from(pipeline
+        const collection = create(new Set([1, null, undefined, 3,]));
+        const called: number[] = Array.from(collection
           .notNullable()
         );
         expect(called[0]).toEqual(1);
@@ -27,18 +27,18 @@ export function testNotNullable(create: <T>(iterable: Iterateable<T>) => IHasNot
           [3, undefined, ],
           [4, 3,],
         ]);
-        const pipeline = create(() => map.values());
-        const called: number[] = Array.from(pipeline
+        const collection = create(() => map.values());
+        const called: number[] = Array.from(collection
           .notNullable()
         );
         expect(called[0]).toEqual(1);
         expect(called[1]).toEqual(3);
       });
       it('generator', () => {
-        const pipeline = create(function * (): Iterable<number | null | undefined> {
+        const collection = create(function * (): Iterable<number | null | undefined> {
           yield 1; yield null; yield undefined; yield 3;
         });
-        const called: number[] = Array.from(pipeline
+        const called: number[] = Array.from(collection
           .notNullable()
         );
         expect(called[0]).toEqual(1);