+++ src/test/zip-short.spec.util.ts
@@ -1,13 +1,13 @@
-import { IHasZipShort } from '../collection.interface';
+import { IHasZip } from '../collection.interface';
 import { Iterateable } from '../types';
 
-export function testZipShort(create: <T>(iterable: Iterateable<T>) => IHasZipShort<T>): void {
-  describe('zipShort(...)', () => {
+export function testZip(create: <T>(iterable: Iterateable<T>) => IHasZip<T>): void {
+  describe('zip(...)', () => {
     describe('should work on', () => {
       it('array', () => {
         const pipeline = create([1, 2, 3,]);
         const called: [number, number][] = Array.from(pipeline
-          .zipShort([4, 5, 6, 7,])
+          .zip([4, 5, 6, 7,])
         );
         expect(called[0]).toEqual([1, 4,]);
         expect(called[1]).toEqual([2, 5,]);
@@ -17,7 +17,7 @@ export function testZipShort(create: <T>(iterable: Iterateable<T>) => IHasZipSho
       it('set', () => {
         const pipeline = create(new Set([1, 2, 3,]));
         const called: [number, number][] = Array.from(pipeline
-          .zipShort([4, 5, 6, 7,])
+          .zip([4, 5, 6, 7,])
         );
         expect(called[0]).toEqual([1, 4,]);
         expect(called[1]).toEqual([2, 5,]);
@@ -28,7 +28,7 @@ export function testZipShort(create: <T>(iterable: Iterateable<T>) => IHasZipSho
         const map = new Map<number, number>([[1, 1,], [2, 2,], [3, 3,],]);
         const pipeline = create(() => map.values());
         const called: [number, number][] = Array.from(pipeline
-          .zipShort([4, 5, 6, 7,])
+          .zip([4, 5, 6, 7,])
         );
         expect(called[0]).toEqual([1, 4,]);
         expect(called[1]).toEqual([2, 5,]);
@@ -43,7 +43,7 @@ export function testZipShort(create: <T>(iterable: Iterateable<T>) => IHasZipSho
         };
         const pipeline = create(generator);
         const called: [number, number][] = Array.from(pipeline
-          .zipShort([4, 5, 6, 7,])
+          .zip([4, 5, 6, 7,])
         );
         expect(called[0]).toEqual([1, 4,]);
         expect(called[1]).toEqual([2, 5,]);